style: use "is False" and "is not False"
Use "is False" and "is not False" in place of "foo == False" or
"foo == True" as recommended by PEP-8.
https://peps.python.org/pep-0008/#programming-recommendations

Gotta say, it makes me really uneasy making changes like this without
any tests. The app is written in an older coding style (cos it's old-ish! :),
which makes me doubly nervous to touch anything. :)